feat(DTFS2-6967): commonise generate and validate userInformation helpers
